+++ db/migrate/20170407114148_increase_position_decimal_precision_to_stories.rb
@@ -1,0 +1,5 @@
+class IncreasePositionDecimalPrecisionToStories < ActiveRecord::Migration
+  def change
+    change_column :stories, :position, :decimal, precision: 25, scale: 20
+  end
+end