+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Redirect automatically to project#index after the user authentication, when there is only one team on enrollments
+- Increase decimal precision from Stories position
 
 ## [1.1.3] - 2017-03-30
 ### Changed